// src/modules/state-manager/types/app-states-names.const.ts
export enum APP_STATES_NAMES {
  TOOLS_META = 'TOOLS_META',
  CHAINS = 'CHAINS',
  PLATFORM_COINS = 'PLATFORM_COINS',
  EXCHANGES_SWAP = 'EXCHANGES_SWAP',
  EXCHANGES_ON_OFF_RAMP = 'EXCHANGES_ON_OFF_RAMP',
  COINGECKO_ID_SYMBOL = 'COINGECKO_ID_SYMBOL',
  SUPPORTED_POOL = 'SUPPORTED_POOL',
  SUPPORTED_POOL_CEX = 'SUPPORTED_POOL_CEX',
  SUPPORTED_RESERVES = 'SUPPORTED_RESERVES',
  SUPPORTED_STAKING_TOKENS = 'SUPPORTED_STAKING_TOKENS',
  SUPPORTED_STABLE_COINS = 'SUPPORTED_STABLE_COINS',
}